+++ tests/test_app/Plugin/TestPlugin/src/Controller/PagesController.php
@@ -20,6 +20,5 @@ class PagesController extends Controller
 {
     public function index()
     {
-
     }
 }